Restructure project to keep backend-classes in commands-package
