+++ bot_commands.py
@@ -1,4 +1,5 @@
-from weather import Weather
+from commands.weather import Weather
+from commands.animals import Animals
 
 
 class BotCommands:
@@ -7,6 +8,10 @@ class BotCommands:
     """
     def __init__(self):
         self.weather = Weather()
+        self.animals = Animals()
 
     def get_weather(self, city):
         return self.weather.fetch_weather(city)
+
+    def get_animal(self, kind):
+        return self.animals.fetch_animalpic(kind)
