+++ commands/data/animals.json
