+++ commands/data/jokes.json
